changed the preconditioners API
